refactor(api): Rename package after app renaming
Renamed api-related folders and packages to match the app name instead of sample.
