+++ app/src/main/java/com/android/gatherly/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package com.android.gatherly.resources
 
 // Like R, but C
 object C {
