+++ app/src/main/java/com/android/gatherly/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.android.gatherly.ui.theme
 
 import androidx.compose.ui.graphics.Color
