+++ app/src/test/java/com/android/gatherly/ExampleRobolectricTest.kt
@@ -1,8 +1,8 @@
-package com.android.sample
+package com.android.gatherly
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.sample.screen.SecondScreen
+import com.android.gatherly.screen.SecondScreen
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import org.junit.Rule
