+++ app/src/test/java/com/android/gatherly/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.gatherly
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
