+++ app/src/test/java/com/android/gatherly/screen/SecondScreen.kt
@@ -1,7 +1,7 @@
-package com.android.sample.screen
+package com.android.gatherly.screen
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.android.sample.resources.C
+import com.android.gatherly.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
